Update NGSIArcGisSinkTest because the test look for linux arcgis library --> https://github.com/telefonicaid/fiware-cygnus/issues/1672
